整体优化代码

+++ COG/Config/Config.cs
@@ -7,13 +7,7 @@ namespace COG.Config;
 
 public class Config
 {
-    public static List<Config> Configs { get; } = new();
-
     public static readonly string DataDirectoryName = $"{Main.DisplayName}_DATA";
-    public string Name { get; }
-    public string Path { get; }
-    public string Text { get; }
-    public Yaml YamlReader { get; private set; }
 
     public Config(string name, string path, string text)
     {
@@ -32,15 +26,18 @@ public class Config
         LoadConfig();
     }
 
+    public static List<Config> Configs { get; } = new();
+    public string Name { get; }
+    public string Path { get; }
+    public string Text { get; }
+    public Yaml YamlReader { get; private set; }
+
     public void LoadConfig(bool replace = false)
     {
         if (!Directory.Exists(DataDirectoryName)) Directory.CreateDirectory(DataDirectoryName);
         if (!File.Exists(Path) || replace)
         {
-            if (replace && File.Exists(Path))
-            {
-                File.Delete(Path);
-            }
+            if (replace && File.Exists(Path)) File.Delete(Path);
             File.WriteAllText(Path, Text, Encoding.Unicode);
         }
